% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data}
\alias{jags_data}
\title{Examples of data transformation uses within the \code{JAGS} code.}
\usage{
jags_data(block)
}
\arguments{
\item{block}{Character string indicating a data generation method or transformation applied to input data.}
}
\description{
\code{jags_data()} prints examples of what can be passed in the data block within the \code{JAGS} code. How the data is
generated can also be specified for  the particular following model, by creating a \code{data}
string before model specification.
}
\section{Details}{

\describe{By specifying \verb{data \{...\}} before \verb{model \{...\}}, the data passed onto the \code{JAGS} model will be transformed or
generated accordingly. There are much more options than the ones shown here, this should be used just as an example.
}
\describe{
\item{Transformation}{Available transformations are "sqrt", "log", "exp", and "sqr". Also "dim" can be used as
an argument. The latter is useful so that the dimension of the response vector doesn't need to be previously
specified.}
\item{Simulated data}{Using "prior" generates true values for the parameters mu and tau from the given prior
specified in the data block. This means the mu and tau are are generated and then passed to the model. Using
"fix" will print a code where mu and tau are the same for both the data and model block. This means the parameters will
be taken as observed values in the model.}
}
}

\examples{
jags_data(block = "sqrt")
jags_data(block = "fix")
}
