% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_examples.R
\name{jags_examples}
\alias{jags_examples}
\title{Coded examples of JAGS models and R runs.}
\usage{
jags_examples(example, runme = T)
}
\arguments{
\item{example}{Character string. Prints example models using \code{JAGS} specific code and parametrization. Options are \code{"model1"} or \verb{"model2}.}

\item{runme}{Logical. If \code{TRUE} the R code to run the \code{JAGS} model will also be displayed. The default value is \code{runme = TRUE}.}
}
\description{
\code{jags_examples()} gives an example of code in \code{JAGS} for model specification, and code that could be passed to \code{run.jags} function.
}
\section{Models}{

\code{example}
\describe{
\item{model 1}{logistic regression for a Bernoulli random variable. Linearity on eta, with a and b normal prior distributions.}
\item{model 2}{logarithmic regression for a Poisson random variable.}
}
\code{runme}
\describe{specification of initial values and distributions for the \code{JAGS} model, as well as \code{run.jags} example code.
}
}

\examples{
\dontrun{
j.ex <- jags_examples("model1")
}

}
\seealso{
\code{\link[runjags]{run.jags}}
}
