% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_model.R
\name{jags_model}
\alias{jags_model}
\title{Returns JAGS code for an specific model}
\usage{
jags_model(likelihood, regression, param = F, reg = F)
}
\arguments{
\item{likelihood}{Likelihood function. It accepts "dnorm", "dbinom", "dpois", and "dbern".}

\item{regression}{A function accepting "linear", "logit", and "log" as part of the likelihood specification.}

\item{param}{Logical with \code{FALSE} as default option. If \code{TRUE}, the prior distribution is previously specified as a character string using the same
parameters as in the \code{regression} argument. If \code{FALSE}, the prior distribution is assumed to be normal.}

\item{reg}{Logical with \code{FALSE} as default option. It only accepts a \code{TRUE} value if \code{regression = "linear"}, and prints the formula for the mean and variance.}
}
\description{
\code{jags_model()} prints the code needed for model specification when using \code{JAGS}. The models and distribution options are quite general,
so the user can grasp the structure of the code. However, this models are widely used without added complexity. Note that \code{JAGS} accepts a
wide range of distributions for the likelihood parameter.
}
\section{Details}{

\describe{There is a lot more flexibility for \code{JAGS} models than the ones specified. However, this is a general
representation of the code one would use. Some other distribution examples can be found in \code{jags_param}.
}
\describe{Note that it only makes sense to use a linear regression with a normal likelihood, since the othe distributions
are discrete distributions.
}
}

\examples{
jags_model("dnorm", "linear", param = FALSE, reg = TRUE)
}
\seealso{
\code{\link{jags_param}}
}
