% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_multivariate.R
\name{jags_multivariate}
\alias{jags_multivariate}
\title{Parametrization in \code{JAGS} for multivariate distribution}
\usage{
jags_multivariate(dist)
}
\arguments{
\item{dist}{Character string indicating the distribution. Argument is "norm" or "multinomial".}
}
\description{
\code{jags_multivariate()} prints the parametrization of the main multivariate distributions built in \code{JAGS}.
}
\examples{
jags_multivariate("multinomial")
}
\seealso{
\code{\link{jags_param}} for univariate distributions.
}
