% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_param.R
\name{jags_param}
\alias{jags_param}
\title{Distribution parametrization in \code{JAGS}}
\usage{
jags_param(dist)
}
\arguments{
\item{dist}{Character string indicating the distribution.}
}
\description{
\code{jags_param()} prints the parametrization of different distributions according to \code{JAGS}.
}
\section{Details}{

\describe{Arguments passed for distribution density may vary from R parametrization, so this function prints
the arguments for a wide range of them.
}
\describe{Distribution that can be specified as the function argument are, "binom", "gamma", "exp", "chisq", "f", "logistic", "lognormal",
"norm", "t", "unif", "bern", "categorical", "pois", "Laplace", "negbin".
}
}

\examples{
\dontrun{
jags_param("f")
}

}
