% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_runjugs.R
\name{jags_runjags}
\alias{jags_runjags}
\title{General code for \code{run.jags} function.}
\usage{
jags_runjags()
}
\description{
\code{jags_runjags()} prints the general arguments passed onto the \code{run.jags} function, and the corresponding class
to which it belongs.
}
\examples{
\dontrun{
jags_runjags()
}

}
